api/requirements.txt
google-generativeai>=0.8.0
groq>=0.5.0
mistralai>=1.0.0
python-dotenv>=1.0.1
